requirements/local.txt
asgiref==3.11.0
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.9
django-allauth==65.13.1
django-browser-reload==1.21.0
django-environ==0.12.0
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.2
